+++ CHANGELOG.rst
@@ -2,7 +2,7 @@
 Changelog
 =========
 
-* Add ``git-conflict-markers``.
+* Add ``git-attributes`` and ``git-conflict-markers`` lexers.
 
 1.1.0 (2023-04-06)
 ------------------